backend/requirements.txt
Flask==2.0.1
flask-bcrypt==1.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
PyJWT==2.3.0
python-dateutil==2.8.2 